move edge id.
